Stopped server from crashing when receiving an invalid token
It will now also show the status of the token in the console. The token's username can only be displayed when the token is expired, because any other exception, e.g. malformed, would mean that the body can't be read correctly. In these cases, the only action is either completely disregarding them (which is the case currently, apart from a console log), or somehow communicating to the client that there is something wrong.